Начал делать окно настроек Каминт перед тем как всё сломать
